refactored location to a context
